--- index.html
@@ -1,45 +1,0 @@
-
-<meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>MarketCycling</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <header>
-        <h1>MarketCycling</h1>
-    </header>
-
-<form id="cadastroCliente">
-    <input type="text" name="nome" placeholder="Nome">
-    <input type="email" name="email" placeholder="E-mail">
-    <input type="tel" name="telefone" placeholder="Telefone">
-    <button type="submit">Cadastrar Cliente</button>
-</form>
-<!DOCTYPE html>
-<html lang="pt-br">
-<head>
-    
-</head>
-<body>
-    
-    
-    <nav class="menu">
-        <ul>
-            <li><a href="/produtos">Produtos</a></li>
-            <li><a href="/vendas">Vendas</a></li>
-            <!-- Adicione mais itens de menu conforme necessário -->
-        </ul>
-    </nav>
-
-    <main>
-        <!-- Conteúdo da página vai aqui -->
-    </main>
-
-    <footer>
-        <p>&copy; 2024 Meu Sistema</p>
-    </footer>
-
-    <script src="scripts.js"></script>
-</body>
-</html>
-
