--- scripts.js
@@ -1,5 +1,0 @@
-document.getElementById('cadastroCliente').addEventListener('submit', function(event) {
-    event.preventDefault();
-    // Lógica para enviar dados do formulário para o servidor e processar a resposta
-    // Exemplo: fetch('url-do-servidor', { method: 'POST', body: new FormData(this) })
-});
