--- src/controllers/userController.js
@@ -1,13 +1,0 @@
-import { createUser, getUserById } from '../models/userModel.js';
-
-// Função para cadastrar um novo usuário
-export async function createUser(nome, email, senha, endereco, telefone, tipo_usuario,) {
-    // Implemente a lógica para criar um novo usuário no banco de dados
-    // Utilize as funções do modelo (userModel.js) para interagir com o banco de dados
-}
-
-// Função para obter um usuário pelo ID
-export async function getUser(id) {
-    // Implemente a lógica para obter um usuário pelo ID no banco de dados
-    // Utilize as funções do modelo (userModel.js) para interagir com o banco de dados
-}
