--- src/models/userModels.js
@@ -1,11 +1,0 @@
-import connection from '../database.js';
-
-// Função para cadastrar um novo usuário
-export async function createUser(nome, email, senha, endereco, telefone, tipo_usuario,) {
-    // Implemente a lógica para inserir um novo usuário no banco de dados
-}
-
-// Função para obter um usuário pelo ID
-export async function getUserById(id) {
-    // Implemente a lógica para obter um usuário pelo ID no banco de dados
-}
