--- src/server.js
@@ -1,16 +1,0 @@
-import dotenv from 'dotenv';
-import express from 'express';
-import cors from 'cors';
-import bodyParser from 'body-parser';
-import routes from './routes.mjs';
-
-dotenv.config({path:'variaveis.env'});
-
-const server = express();
-server.use(cors());
-server.use(bodyParser.urlencoded({extended: false}));
-
-server.listen(process.env.PORT, () => {
-    console.log(`Servidor rodando em: http://localhost:${process.env.PORT}`);
-});
-
