--- style.css
@@ -1,46 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-}
-
-header, nav, main, footer {
-    padding: 20px;
-}
-
-.menu {
-    background-color: #333;
-}
-
-.menu ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-.menu ul li {
-    display: inline;
-    margin-right: 20px;
-}
-
-.menu ul li a {
-    color: #fff;
-    text-decoration: none;
-}
-
-input[type="text"], input[type="email"], input[type="tel"], button {
-    margin-bottom: 10px;
-    display: block;
-}
-
-button {
-    background-color: #333;
-    color: #fff;
-    border: none;
-    padding: 10px 20px;
-    cursor: pointer;
-}
-
-button:hover {
-    background-color: #555;
-}
